+++ semanticcache/types.py
@@ -16,6 +16,7 @@ CacheSource = Literal[
 
 
 EmbedderType = Literal[
+    "custom",
     "huggingface",
     "openai",
     "cohere",